move store methods to keeper.go
